+++ kriolos-opos-app/src/main/java/com/openbravo/pos/imports/StockQtyImport.java
@@ -26,6 +26,7 @@ import com.openbravo.pos.ticket.ProductInfoExt;
 
 import com.openbravo.pos.inventory.InventoryService;
 import com.openbravo.pos.inventory.InventoryServiceImpl;
+import com.openbravo.pos.pim.DataLogicPIM;
 
 import javax.swing.*;
 import javax.swing.event.DocumentEvent;
@@ -58,6 +59,8 @@ public class StockQtyImport extends JPanel implements JPanelView {
   private DataLogicSystem m_dlSystem;
   private DataLogicImport m_dlImport;
   private InventoryService invService;
+  private DataLogicPIM dataLogicPIM;
+
   private ProductInfoExt prodInfo;
   private ProductStock prodStock;
 
@@ -93,6 +96,9 @@ public class StockQtyImport extends JPanel implements JPanelView {
     initComponents();
 
     dbSession = appView.getSession();
+    
+    dataLogicPIM = new DataLogicPIM(); 
+    dataLogicPIM.init(dbSession);
 
     // Set db tables
     m_dlSales = new DataLogicSales();
@@ -332,7 +338,7 @@ public class StockQtyImport extends JPanel implements JPanelView {
 
       String sCode = products.get(0);
 
-      prodInfo = m_dlSales.getProductInfoByCode(sCode);
+      prodInfo = dataLogicPIM.getProductInfoByCode(sCode);
 
       if (prodInfo != null) {
         prodStock = m_dlSales.getProductStockState(prodInfo.getID(), m_sInventoryLocation);
@@ -361,7 +367,7 @@ public class StockQtyImport extends JPanel implements JPanelView {
 
       String sCode = products.get(0);
 
-      prodInfo = m_dlSales.getProductInfoByCode(sCode);
+      prodInfo = dataLogicPIM.getProductInfoByCode(sCode);
 
       if (prodInfo != null) {
         prodStock = m_dlSales.getProductStockState(prodInfo.getID(), m_sInventoryLocation);
@@ -492,292 +498,254 @@ public class StockQtyImport extends JPanel implements JPanelView {
    * regenerated by the Form Editor.
    */
   // <editor-fold defaultstate="collapsed" desc="Generated
-  // Code">//GEN-BEGIN:initComponents
-  private void initComponents() {
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
 
-    jFileChooserPanel = new javax.swing.JPanel();
-    jLabel1 = new javax.swing.JLabel();
-    jFileName = new javax.swing.JTextField();
-    jLabel18 = new javax.swing.JLabel();
-    jbtnFileChoose = new javax.swing.JButton();
-    jComboSeparator = new javax.swing.JComboBox();
-    webPBar = new javax.swing.JProgressBar();
-    jPanel2 = new javax.swing.JPanel();
-    jLabel9 = new javax.swing.JLabel();
-    jTextUpdates = new javax.swing.JLabel();
-    jTextRecords = new javax.swing.JTextField();
-    jTextUpdate = new javax.swing.JTextField();
-    jFileRead = new javax.swing.JButton();
-    m_jLocation = new javax.swing.JLabel();
-    jImport = new javax.swing.JButton();
-    jbtnReset = new javax.swing.JButton();
-    jLabel2 = new javax.swing.JLabel();
+        jFileChooserPanel = new javax.swing.JPanel();
+        jLabel1 = new javax.swing.JLabel();
+        jFileName = new javax.swing.JTextField();
+        jLabel18 = new javax.swing.JLabel();
+        jbtnFileChoose = new javax.swing.JButton();
+        jComboSeparator = new javax.swing.JComboBox();
+        webPBar = new javax.swing.JProgressBar();
+        jPanel2 = new javax.swing.JPanel();
+        jLabel9 = new javax.swing.JLabel();
+        jTextUpdates = new javax.swing.JLabel();
+        jTextRecords = new javax.swing.JTextField();
+        jTextUpdate = new javax.swing.JTextField();
+        jFileRead = new javax.swing.JButton();
+        m_jLocation = new javax.swing.JLabel();
+        jImport = new javax.swing.JButton();
+        jbtnReset = new javax.swing.JButton();
+        jLabel2 = new javax.swing.JLabel();
 
-    setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
-    setMaximumSize(new java.awt.Dimension(700, 500));
-    setMinimumSize(new java.awt.Dimension(700, 450));
-    setOpaque(false);
-    setPreferredSize(new java.awt.Dimension(700, 500));
+        setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
+        setMaximumSize(new java.awt.Dimension(700, 500));
+        setMinimumSize(new java.awt.Dimension(700, 450));
+        setOpaque(false);
+        setPreferredSize(new java.awt.Dimension(700, 500));
 
-    jLabel1.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
-    java.util.ResourceBundle bundle = java.util.ResourceBundle.getBundle("pos_messages"); // NOI18N
-    jLabel1.setText(bundle.getString("label.csvfile")); // NOI18N
-    jLabel1.setMaximumSize(new java.awt.Dimension(23, 45));
-    jLabel1.setPreferredSize(new java.awt.Dimension(100, 45));
+        jLabel1.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
+        java.util.ResourceBundle bundle = java.util.ResourceBundle.getBundle("pos_messages"); // NOI18N
+        jLabel1.setText(bundle.getString("label.csvfile")); // NOI18N
+        jLabel1.setMaximumSize(new java.awt.Dimension(23, 45));
+        jLabel1.setPreferredSize(new java.awt.Dimension(100, 45));
 
-    jFileName.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
-    jFileName.setPreferredSize(new java.awt.Dimension(300, 45));
-    jFileName.addActionListener(new java.awt.event.ActionListener() {
-      public void actionPerformed(java.awt.event.ActionEvent evt) {
-        jFileNameActionPerformed(evt);
-      }
-    });
+        jFileName.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
+        jFileName.setPreferredSize(new java.awt.Dimension(300, 45));
+        jFileName.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jFileNameActionPerformed(evt);
+            }
+        });
 
-    jLabel18.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
-    jLabel18.setText(bundle.getString("label.csvdelimit")); // NOI18N
-    jLabel18.setPreferredSize(new java.awt.Dimension(100, 30));
+        jLabel18.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
+        jLabel18.setText(bundle.getString("label.csvdelimit")); // NOI18N
+        jLabel18.setPreferredSize(new java.awt.Dimension(100, 30));
 
-    jbtnFileChoose.setIcon(new javax.swing.ImageIcon(getClass().getResource("/com/openbravo/images/fileopen.png"))); // NOI18N
-    jbtnFileChoose.setMaximumSize(new java.awt.Dimension(64, 32));
-    jbtnFileChoose.setMinimumSize(new java.awt.Dimension(64, 32));
-    jbtnFileChoose.setPreferredSize(new java.awt.Dimension(80, 45));
-    jbtnFileChoose.addActionListener(new java.awt.event.ActionListener() {
-      public void actionPerformed(java.awt.event.ActionEvent evt) {
-        jbtnFileChooseActionPerformed(evt);
-      }
-    });
+        jbtnFileChoose.setIcon(new javax.swing.ImageIcon(getClass().getResource("/com/openbravo/images/fileopen.png"))); // NOI18N
+        jbtnFileChoose.setMaximumSize(new java.awt.Dimension(64, 32));
+        jbtnFileChoose.setMinimumSize(new java.awt.Dimension(64, 32));
+        jbtnFileChoose.setPreferredSize(new java.awt.Dimension(80, 45));
+        jbtnFileChoose.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jbtnFileChooseActionPerformed(evt);
+            }
+        });
 
-    jComboSeparator.setFont(new java.awt.Font("Arial", 0, 12)); // NOI18N
-    jComboSeparator.setMaximumSize(new java.awt.Dimension(50, 45));
-    jComboSeparator.setPreferredSize(new java.awt.Dimension(50, 45));
+        jComboSeparator.setFont(new java.awt.Font("Arial", 0, 12)); // NOI18N
+        jComboSeparator.setMaximumSize(new java.awt.Dimension(50, 45));
+        jComboSeparator.setPreferredSize(new java.awt.Dimension(50, 45));
 
-    webPBar.setFont(new java.awt.Font("Arial", 0, 13)); // NOI18N
-    webPBar.setMaximumSize(new java.awt.Dimension(240, 45));
-    webPBar.setMinimumSize(new java.awt.Dimension(10, 45));
-    webPBar.setPreferredSize(new java.awt.Dimension(240, 45));
+        webPBar.setFont(new java.awt.Font("Arial", 0, 13)); // NOI18N
+        webPBar.setMaximumSize(new java.awt.Dimension(240, 45));
+        webPBar.setMinimumSize(new java.awt.Dimension(10, 45));
+        webPBar.setPreferredSize(new java.awt.Dimension(240, 45));
 
-    jPanel2.setBorder(javax.swing.BorderFactory.createTitledBorder(
-        new javax.swing.border.LineBorder(new java.awt.Color(153, 153, 153), 1, true),
-        bundle.getString("title.CSVImport"), javax.swing.border.TitledBorder.DEFAULT_JUSTIFICATION,
-        javax.swing.border.TitledBorder.DEFAULT_POSITION, new java.awt.Font("Arial", 1, 14),
-        new java.awt.Color(102, 102, 102))); // NOI18N
+        jPanel2.setBorder(javax.swing.BorderFactory.createTitledBorder(new javax.swing.border.LineBorder(new java.awt.Color(153, 153, 153), 1, true), bundle.getString("title.CSVImport"), javax.swing.border.TitledBorder.DEFAULT_JUSTIFICATION, javax.swing.border.TitledBorder.DEFAULT_POSITION, new java.awt.Font("Arial", 1, 14), new java.awt.Color(102, 102, 102))); // NOI18N
 
-    jLabel9.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
-    jLabel9.setText(bundle.getString("label.csvrecordsfound")); // NOI18N
-    jLabel9.setPreferredSize(new java.awt.Dimension(150, 30));
+        jLabel9.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
+        jLabel9.setText(bundle.getString("label.csvrecordsfound")); // NOI18N
+        jLabel9.setPreferredSize(new java.awt.Dimension(150, 30));
 
-    jTextUpdates.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
-    jTextUpdates.setText(bundle.getString("label.csvchanged")); // NOI18N
-    jTextUpdates.setPreferredSize(new java.awt.Dimension(150, 30));
+        jTextUpdates.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
+        jTextUpdates.setText(bundle.getString("label.csvchanged")); // NOI18N
+        jTextUpdates.setPreferredSize(new java.awt.Dimension(150, 30));
 
-    jTextRecords.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
-    jTextRecords.setForeground(new java.awt.Color(102, 102, 102));
-    jTextRecords.setHorizontalAlignment(javax.swing.JTextField.RIGHT);
-    jTextRecords.setBorder(null);
-    jTextRecords.setDisabledTextColor(new java.awt.Color(0, 0, 0));
-    jTextRecords.setEnabled(false);
-    jTextRecords.setPreferredSize(new java.awt.Dimension(100, 30));
+        jTextRecords.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
+        jTextRecords.setForeground(new java.awt.Color(102, 102, 102));
+        jTextRecords.setHorizontalAlignment(javax.swing.JTextField.RIGHT);
+        jTextRecords.setBorder(null);
+        jTextRecords.setDisabledTextColor(new java.awt.Color(0, 0, 0));
+        jTextRecords.setEnabled(false);
+        jTextRecords.setPreferredSize(new java.awt.Dimension(100, 30));
 
-    jTextUpdate.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
-    jTextUpdate.setForeground(new java.awt.Color(102, 102, 102));
-    jTextUpdate.setHorizontalAlignment(javax.swing.JTextField.RIGHT);
-    jTextUpdate.setBorder(null);
-    jTextUpdate.setDisabledTextColor(new java.awt.Color(0, 0, 0));
-    jTextUpdate.setEnabled(false);
-    jTextUpdate.setPreferredSize(new java.awt.Dimension(100, 30));
+        jTextUpdate.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
+        jTextUpdate.setForeground(new java.awt.Color(102, 102, 102));
+        jTextUpdate.setHorizontalAlignment(javax.swing.JTextField.RIGHT);
+        jTextUpdate.setBorder(null);
+        jTextUpdate.setDisabledTextColor(new java.awt.Color(0, 0, 0));
+        jTextUpdate.setEnabled(false);
+        jTextUpdate.setPreferredSize(new java.awt.Dimension(100, 30));
 
-    javax.swing.GroupLayout jPanel2Layout = new javax.swing.GroupLayout(jPanel2);
-    jPanel2.setLayout(jPanel2Layout);
-    jPanel2Layout.setHorizontalGroup(
-        jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+        javax.swing.GroupLayout jPanel2Layout = new javax.swing.GroupLayout(jPanel2);
+        jPanel2.setLayout(jPanel2Layout);
+        jPanel2Layout.setHorizontalGroup(
+            jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(jPanel2Layout.createSequentialGroup()
                 .addContainerGap()
-                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
                     .addGroup(jPanel2Layout.createSequentialGroup()
-                        .addComponent(jLabel9, javax.swing.GroupLayout.PREFERRED_SIZE,
-                            javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addComponent(jLabel9, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jTextRecords, javax.swing.GroupLayout.PREFERRED_SIZE,
-                            javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addComponent(jTextRecords, javax.swing.GroupLayout.PREFERRED_SIZE, 160, javax.swing.GroupLayout.PREFERRED_SIZE))
                     .addGroup(jPanel2Layout.createSequentialGroup()
-                        .addComponent(jTextUpdates, javax.swing.GroupLayout.DEFAULT_SIZE,
-                            javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addComponent(jTextUpdates, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jTextUpdate, javax.swing.GroupLayout.PREFERRED_SIZE,
-                            javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)));
-    jPanel2Layout.setVerticalGroup(
-        jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(jTextUpdate, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)))
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+        );
+        jPanel2Layout.setVerticalGroup(
+            jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(jPanel2Layout.createSequentialGroup()
                 .addGap(0, 0, 0)
                 .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel9, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE,
-                        javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jTextRecords, javax.swing.GroupLayout.PREFERRED_SIZE,
-                        javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addComponent(jLabel9, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jTextRecords, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jTextUpdates, javax.swing.GroupLayout.PREFERRED_SIZE,
-                        javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jTextUpdate, javax.swing.GroupLayout.PREFERRED_SIZE,
-                        javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap()));
+                    .addComponent(jTextUpdates, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jTextUpdate, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addContainerGap())
+        );
 
-    javax.swing.GroupLayout jFileChooserPanelLayout = new javax.swing.GroupLayout(jFileChooserPanel);
-    jFileChooserPanel.setLayout(jFileChooserPanelLayout);
-    jFileChooserPanelLayout.setHorizontalGroup(
-        jFileChooserPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+        javax.swing.GroupLayout jFileChooserPanelLayout = new javax.swing.GroupLayout(jFileChooserPanel);
+        jFileChooserPanel.setLayout(jFileChooserPanelLayout);
+        jFileChooserPanelLayout.setHorizontalGroup(
+            jFileChooserPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(jFileChooserPanelLayout.createSequentialGroup()
-                .addGroup(jFileChooserPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jFileChooserPanelLayout.createSequentialGroup()
-                        .addComponent(jLabel1, javax.swing.GroupLayout.DEFAULT_SIZE,
-                            javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jFileName, javax.swing.GroupLayout.PREFERRED_SIZE,
-                            javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jbtnFileChoose, javax.swing.GroupLayout.PREFERRED_SIZE,
-                            javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING,
-                        jFileChooserPanelLayout.createSequentialGroup()
-                            .addComponent(jLabel18, javax.swing.GroupLayout.PREFERRED_SIZE,
-                                javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                            .addComponent(jComboSeparator, javax.swing.GroupLayout.PREFERRED_SIZE,
-                                javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED,
-                                javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                            .addGroup(jFileChooserPanelLayout
-                                .createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                                .addComponent(jPanel2, javax.swing.GroupLayout.Alignment.TRAILING,
-                                    javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE,
-                                    Short.MAX_VALUE)
-                                .addComponent(webPBar, javax.swing.GroupLayout.Alignment.TRAILING,
-                                    javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE,
-                                    Short.MAX_VALUE))))
-                .addContainerGap()));
-    jFileChooserPanelLayout.setVerticalGroup(
-        jFileChooserPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jFileChooserPanelLayout.createSequentialGroup()
-                .addGroup(jFileChooserPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jbtnFileChoose, javax.swing.GroupLayout.Alignment.TRAILING,
-                        javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE,
-                        javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addGroup(jFileChooserPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                        .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE,
-                            javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addComponent(jFileName, javax.swing.GroupLayout.PREFERRED_SIZE,
-                            javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                .addGroup(jFileChooserPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                    .addComponent(jLabel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(jLabel18, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addGroup(jFileChooserPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jComboSeparator, javax.swing.GroupLayout.DEFAULT_SIZE,
-                        javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addGroup(jFileChooserPanelLayout.createSequentialGroup()
+                        .addComponent(jComboSeparator, javax.swing.GroupLayout.PREFERRED_SIZE, 66, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(webPBar, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jFileChooserPanelLayout.createSequentialGroup()
+                        .addComponent(jFileName, javax.swing.GroupLayout.DEFAULT_SIZE, 478, Short.MAX_VALUE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jbtnFileChoose, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addComponent(jPanel2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap())
+        );
+        jFileChooserPanelLayout.setVerticalGroup(
+            jFileChooserPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jFileChooserPanelLayout.createSequentialGroup()
+                .addGroup(jFileChooserPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jbtnFileChoose, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addGroup(jFileChooserPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                        .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addComponent(jFileName, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(jFileChooserPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jComboSeparator, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                     .addGroup(jFileChooserPanelLayout.createSequentialGroup()
                         .addGap(0, 0, Short.MAX_VALUE)
-                        .addComponent(jLabel18, javax.swing.GroupLayout.PREFERRED_SIZE,
-                            javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addComponent(jLabel18, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                         .addGap(2, 2, 2))
                     .addGroup(jFileChooserPanelLayout.createSequentialGroup()
-                        .addComponent(webPBar, javax.swing.GroupLayout.PREFERRED_SIZE,
-                            javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addComponent(webPBar, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                         .addGap(0, 0, Short.MAX_VALUE)))
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE,
-                    javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap()));
+                .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap())
+        );
 
-    jFileRead.setFont(new java.awt.Font("Arial", 0, 12)); // NOI18N
-    jFileRead.setText(bundle.getString("label.csvread")); // NOI18N
-    jFileRead.setEnabled(false);
-    jFileRead.setPreferredSize(new java.awt.Dimension(85, 45));
-    jFileRead.addActionListener(new java.awt.event.ActionListener() {
-      public void actionPerformed(java.awt.event.ActionEvent evt) {
-        jFileReadActionPerformed(evt);
-      }
-    });
+        jFileRead.setFont(new java.awt.Font("Arial", 0, 12)); // NOI18N
+        jFileRead.setText(bundle.getString("label.csvread")); // NOI18N
+        jFileRead.setEnabled(false);
+        jFileRead.setPreferredSize(new java.awt.Dimension(85, 45));
+        jFileRead.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jFileReadActionPerformed(evt);
+            }
+        });
 
-    m_jLocation.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
-    m_jLocation.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-    m_jLocation.setText(bundle.getString("label.location")); // NOI18N
-    m_jLocation.setPreferredSize(new java.awt.Dimension(100, 45));
+        m_jLocation.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
+        m_jLocation.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        m_jLocation.setText(bundle.getString("label.location")); // NOI18N
+        m_jLocation.setPreferredSize(new java.awt.Dimension(100, 45));
 
-    jImport.setFont(new java.awt.Font("Arial", 0, 12)); // NOI18N
-    jImport.setText(bundle.getString("label.csvimpostbtn")); // NOI18N
-    jImport.setEnabled(false);
-    jImport.setPreferredSize(new java.awt.Dimension(85, 45));
-    jImport.addActionListener(new java.awt.event.ActionListener() {
-      public void actionPerformed(java.awt.event.ActionEvent evt) {
-        jImportActionPerformed(evt);
-      }
-    });
+        jImport.setFont(new java.awt.Font("Arial", 0, 12)); // NOI18N
+        jImport.setText(bundle.getString("label.csvimpostbtn")); // NOI18N
+        jImport.setEnabled(false);
+        jImport.setPreferredSize(new java.awt.Dimension(85, 45));
+        jImport.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jImportActionPerformed(evt);
+            }
+        });
 
-    jbtnReset.setFont(new java.awt.Font("Arial", 0, 12)); // NOI18N
-    jbtnReset.setText(bundle.getString("button.reset")); // NOI18N
-    jbtnReset.setPreferredSize(new java.awt.Dimension(85, 45));
-    jbtnReset.addActionListener(new java.awt.event.ActionListener() {
-      public void actionPerformed(java.awt.event.ActionEvent evt) {
-        jbtnResetActionPerformed(evt);
-      }
-    });
+        jbtnReset.setFont(new java.awt.Font("Arial", 0, 12)); // NOI18N
+        jbtnReset.setText(bundle.getString("button.reset")); // NOI18N
+        jbtnReset.setPreferredSize(new java.awt.Dimension(85, 45));
+        jbtnReset.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jbtnResetActionPerformed(evt);
+            }
+        });
 
-    jLabel2.setBackground(new java.awt.Color(255, 255, 255));
-    jLabel2.setFont(new java.awt.Font("Arial", 0, 12)); // NOI18N
-    jLabel2.setText(bundle.getString("label.csvnotice")); // NOI18N
-    jLabel2.setVerticalAlignment(javax.swing.SwingConstants.TOP);
-    jLabel2.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(153, 153, 153)));
-    jLabel2.setMaximumSize(new java.awt.Dimension(680, 120));
-    jLabel2.setOpaque(true);
-    jLabel2.setPreferredSize(new java.awt.Dimension(680, 120));
+        jLabel2.setBackground(new java.awt.Color(255, 255, 255));
+        jLabel2.setFont(new java.awt.Font("Arial", 0, 12)); // NOI18N
+        jLabel2.setText(bundle.getString("label.csvnotice")); // NOI18N
+        jLabel2.setVerticalAlignment(javax.swing.SwingConstants.TOP);
+        jLabel2.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(153, 153, 153)));
+        jLabel2.setMaximumSize(new java.awt.Dimension(680, 120));
+        jLabel2.setOpaque(true);
+        jLabel2.setPreferredSize(new java.awt.Dimension(680, 120));
 
-    javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
-    this.setLayout(layout);
-    layout.setHorizontalGroup(
-        layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
+        this.setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(layout.createSequentialGroup()
-                        .addContainerGap()
-                        .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE,
-                            javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(layout.createSequentialGroup()
-                        .addGap(10, 10, 10)
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                            .addComponent(m_jLocation, javax.swing.GroupLayout.PREFERRED_SIZE, 295,
-                                javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(jFileChooserPanel, javax.swing.GroupLayout.PREFERRED_SIZE,
-                                javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addGroup(layout.createSequentialGroup()
-                                .addComponent(jFileRead, javax.swing.GroupLayout.PREFERRED_SIZE,
-                                    javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addComponent(jImport, javax.swing.GroupLayout.PREFERRED_SIZE,
-                                    javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addComponent(jbtnReset, javax.swing.GroupLayout.PREFERRED_SIZE,
-                                    javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))))
-                .addGap(378, 378, 378)));
-    layout.setVerticalGroup(
-        layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                    .addComponent(jFileChooserPanel, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(layout.createSequentialGroup()
+                            .addContainerGap()
+                            .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addGroup(layout.createSequentialGroup()
+                            .addGap(213, 213, 213)
+                            .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                                .addComponent(m_jLocation, javax.swing.GroupLayout.PREFERRED_SIZE, 295, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGroup(layout.createSequentialGroup()
+                                    .addComponent(jFileRead, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                    .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                    .addComponent(jImport, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                    .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                    .addComponent(jbtnReset, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))))))
+                .addGap(378, 378, 378))
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
                 .addGap(19, 19, 19)
-                .addComponent(jFileChooserPanel, javax.swing.GroupLayout.PREFERRED_SIZE,
-                    javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(jFileChooserPanel, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(m_jLocation, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE,
-                    javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(m_jLocation, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.CENTER)
-                    .addComponent(jFileRead, javax.swing.GroupLayout.PREFERRED_SIZE,
-                        javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jImport, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE,
-                        javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jbtnReset, javax.swing.GroupLayout.PREFERRED_SIZE,
-                        javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addComponent(jFileRead, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jImport, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jbtnReset, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
                 .addGap(18, 18, 18)
-                .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, 133,
-                    javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(14, Short.MAX_VALUE)));
-  }// </editor-fold>//GEN-END:initComponents
+                .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, 133, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(21, Short.MAX_VALUE))
+        );
+    }// </editor-fold>//GEN-END:initComponents
 
   private void jFileReadActionPerformed(java.awt.event.ActionEvent evt) {// GEN-FIRST:event_jFileReadActionPerformed
     try {
@@ -848,23 +816,23 @@ public class StockQtyImport extends JPanel implements JPanelView {
     thread2.start();
   }// GEN-LAST:event_jImportActionPerformed
 
-  // Variables declaration - do not modify//GEN-BEGIN:variables
-  private javax.swing.JComboBox jComboSeparator;
-  private javax.swing.JPanel jFileChooserPanel;
-  private javax.swing.JTextField jFileName;
-  private javax.swing.JButton jFileRead;
-  private javax.swing.JButton jImport;
-  private javax.swing.JLabel jLabel1;
-  private javax.swing.JLabel jLabel18;
-  private javax.swing.JLabel jLabel2;
-  private javax.swing.JLabel jLabel9;
-  private javax.swing.JPanel jPanel2;
-  private javax.swing.JTextField jTextRecords;
-  private javax.swing.JTextField jTextUpdate;
-  private javax.swing.JLabel jTextUpdates;
-  private javax.swing.JButton jbtnFileChoose;
-  private javax.swing.JButton jbtnReset;
-  private javax.swing.JLabel m_jLocation;
-  private javax.swing.JProgressBar webPBar;
-  // End of variables declaration//GEN-END:variables
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JComboBox jComboSeparator;
+    private javax.swing.JPanel jFileChooserPanel;
+    private javax.swing.JTextField jFileName;
+    private javax.swing.JButton jFileRead;
+    private javax.swing.JButton jImport;
+    private javax.swing.JLabel jLabel1;
+    private javax.swing.JLabel jLabel18;
+    private javax.swing.JLabel jLabel2;
+    private javax.swing.JLabel jLabel9;
+    private javax.swing.JPanel jPanel2;
+    private javax.swing.JTextField jTextRecords;
+    private javax.swing.JTextField jTextUpdate;
+    private javax.swing.JLabel jTextUpdates;
+    private javax.swing.JButton jbtnFileChoose;
+    private javax.swing.JButton jbtnReset;
+    private javax.swing.JLabel m_jLocation;
+    private javax.swing.JProgressBar webPBar;
+    // End of variables declaration//GEN-END:variables
 }
