+++ kriolos-opos-app/src/main/java/com/openbravo/pos/inventory/CategoriesEditor.form
@@ -55,7 +55,7 @@
                               <Group type="103" groupAlignment="1" max="-2" attributes="0">
                                   <Component id="jLblCatOrder" max="32767" attributes="0"/>
                                   <Component id="jLblInCatalog" max="32767" attributes="0"/>
-                                  <Component id="jLblCatShowName" max="32767" attributes="0"/>
+                                  <Component id="jLblCatShowName" pref="152" max="32767" attributes="0"/>
                                   <Component id="jLblCatColor" max="32767" attributes="0"/>
                               </Group>
                               <EmptySpace min="0" pref="0" max="-2" attributes="0"/>
