+++ kriolos-opos-app/src/main/java/com/openbravo/pos/panels/BundleFilter.java
@@ -23,6 +23,7 @@ import com.openbravo.data.loader.SerializerWriteString;
 import com.openbravo.pos.forms.AppLocal;
 import com.openbravo.pos.forms.AppView;
 import com.openbravo.pos.forms.DataLogicSales;
+import com.openbravo.pos.pim.DataLogicPIM;
 import com.openbravo.pos.reports.ReportEditorCreator;
 import com.openbravo.pos.ticket.ProductInfoExt;
 import java.awt.Component;
@@ -39,7 +40,7 @@ import javax.swing.event.EventListenerList;
 public class BundleFilter extends javax.swing.JPanel implements ReportEditorCreator {
 
     private ProductInfoExt product;
-    private DataLogicSales m_dlSales;
+    private DataLogicPIM dataLogicPIM;
     private AppView appView;
     
     protected EventListenerList listeners = new EventListenerList();
@@ -53,7 +54,7 @@ public class BundleFilter extends javax.swing.JPanel implements ReportEditorCrea
     public void init(AppView app) {   
            
         this.appView = app;
-        m_dlSales = (DataLogicSales) app.getBean("com.openbravo.pos.forms.DataLogicSales");
+        this.dataLogicPIM = (DataLogicPIM) app.getBean("com.openbravo.pos.pim.DataLogicPIM");
     }
 
     @Override
@@ -122,7 +123,7 @@ public class BundleFilter extends javax.swing.JPanel implements ReportEditorCrea
 
     private void assignProductByCode() {
         try {
-            ProductInfoExt prod = m_dlSales.getProductInfoByCode(m_jBarcode1.getText());
+            ProductInfoExt prod = dataLogicPIM.getProductInfoByCode(m_jBarcode1.getText());
             if (prod == null) {
                 Toolkit.getDefaultToolkit().beep();
             }
@@ -136,7 +137,7 @@ public class BundleFilter extends javax.swing.JPanel implements ReportEditorCrea
 
     private void assignProductByReference() {
         try {
-            ProductInfoExt prod = m_dlSales.getProductInfoByReference(m_jReference1.getText());
+            ProductInfoExt prod = dataLogicPIM.getProductInfoByReference(m_jReference1.getText());
             if (prod == null) {
                 Toolkit.getDefaultToolkit().beep();
             }
